backend/requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
click==8.0.3
Flask==2.0.2
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.0
Pillow==9.0.0
pycodestyle==2.8.0
pytz==2021.3
six==1.16.0
toml==0.10.2
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
Werkzeug==2.0.2
